Implement main functionality
Add a main to the program which handles all of the event broker functionality.

Also change the period04 functionality to work with functions instead of a class, in order to allow it to work better with the event broker.

Additionally change the period04 functionality to work with a single temporary directory. Also add the temporary directory to the gitignore file.